added component for placed-item-list
