+++ PokeOneWeb/Frontend/pokeoneweb/src/app/core/components/placed-item-list/core/placed-item-list-column.enum.ts
