+++ PokeOneWeb/Frontend/pokeoneweb/src/app/core/components/placed-item-list/core/placed-item-list-sort.service.ts
@@ -5,7 +5,7 @@ import { PlacedItemListColumn } from './placed-item-list-column.enum';
 @Injectable({
     providedIn: 'root'
 })
-export class ItemDetailSortService {
+export class PlacedItemListSortService {
     public sortPlacedItems(
         models: IPlacedItemModel[],
         sortColumn: PlacedItemListColumn,
