+++ PokeOneWeb/Frontend/pokeoneweb/src/app/core/components/placed-item-list/core/placed-item-list.model.ts
@@ -1,0 +1,9 @@
+import { IItemModel } from 'src/app/core/models/item.model';
+import { PlacedItemListColumn } from './placed-item-list-column.enum';
+
+export class PlacedItemListModel {
+    public placedItemsSortedByColumn: PlacedItemListColumn = PlacedItemListColumn.Location;
+    public placedItemsSortDirection = 1;
+
+    public item: IItemModel | null = null;
+}
