+++ PokeOneWeb/Frontend/pokeoneweb/src/app/core/components/placed-item-list/placed-item-list.component.ts
@@ -1,4 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
+import { IItemModel, ItemModel } from '../../models/item.model';
+import { PlacedItemListColumn } from './core/placed-item-list-column.enum';
+import { PlacedItemListSortService } from './core/placed-item-list-sort.service';
+import { PlacedItemListModel } from './core/placed-item-list.model';
 
 @Component({
   selector: 'app-placed-item-list',
@@ -6,10 +10,50 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./placed-item-list.component.scss']
 })
 export class PlacedItemListComponent implements OnInit {
+  @Input() item : IItemModel = new ItemModel;
 
-  constructor() { }
+  public model : PlacedItemListModel = new PlacedItemListModel();
+  public PlacedItemListColumn = PlacedItemListColumn;
+
+  constructor(
+    private sortService: PlacedItemListSortService
+  ) { }
 
   ngOnInit(): void {
+    this.model.item = this.item;
+    this.applyInitialSorting();
   }
 
+  public sortPlacedItems(sortColumn: PlacedItemListColumn, sortDirection: number) {
+    if (!this.model.item) return;
+
+    this.model.placedItemsSortedByColumn = sortColumn;
+    this.model.placedItemsSortDirection = sortDirection;
+
+    this.model.item.placedItems = this.sortService.sortPlacedItems(
+        this.model.item.placedItems,
+        sortColumn,
+        sortDirection
+    );
+  }
+
+  public getPlacedItemSortButtonClass(sortColumn: PlacedItemListColumn, sortDirection: number): string {
+      if (
+          this.model.placedItemsSortedByColumn === sortColumn &&
+          this.model.placedItemsSortDirection === sortDirection
+      ) {
+          return 'sorted';
+      }
+      return 'unsorted';
+  }
+
+  private applyInitialSorting() {
+    if (!this.model.item) return;
+
+    this.model.item.placedItems = this.sortService.sortPlacedItems(
+        this.model.item.placedItems,
+        PlacedItemListColumn.Location,
+        1
+    );
+  }
 }
